refactor: move Role model in permissions app
