+++ backend/backend/apps/permissions/admin.py
@@ -1,6 +1,11 @@
 from django.contrib import admin
 
-from backend.apps.permissions.models import Permission
+from backend.apps.permissions.models import Permission, Role
+
+
+@admin.register(Role)
+class RoleAdmin(admin.ModelAdmin):
+    pass
 
 
 @admin.register(Permission)
