+++ backend/backend/apps/permissions/models.py
@@ -1,6 +1,18 @@
 from django.db import models
 
 
+class Role(models.Model):
+    display_name = models.CharField(max_length=100)
+    slug = models.SlugField(unique=True, max_length=50)
+    description = models.CharField(blank=True, null=True, max_length=500)
+
+    is_active = models.BooleanField(default=True)
+    permissions = models.ManyToManyField("Permission", related_name="roles", blank=True)
+
+    def __str__(self) -> str:
+        return self.display_name
+
+
 class Permission(models.Model):
     display_name = models.CharField(max_length=100)
     code = models.SlugField(unique=True, max_length=50)
